实现了注册的邮箱验证，同时把验证码保存在Redis中

+++ src/main/java/com/code/pojo/TStudent.java
@@ -1,12 +1,12 @@
 package com.code.pojo;
 
 import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.Version;
 import com.baomidou.mybatisplus.annotation.TableId;
-import java.io.Serializable;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
+
+import java.io.Serializable;
 
 /**
  * <p>
@@ -24,7 +24,7 @@ public class TStudent implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "student_id", type = IdType.AUTO)
-    private Integer studentId;
+    private Long studentId;
 
     private String studentNo;
